added basic styling for job tile
