+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "faker": "^4.1.0",
     "firebase": "^5.9.2",
+    "node-sass": "^4.11.0",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-router-dom": "^5.0.0",
