+++ src/App.css
@@ -7,7 +7,7 @@
 }
 
 .App {
-  text-align: center;
+  /*text-align: center;*/
 }
 
 .App-header {
