+++ src/components/JobsContainer/JobTile/JobTile.jsx
@@ -1,0 +1,49 @@
+import React, { Component } from 'react';
+// import Firebase from '../../../services/firebase';
+import './JobTile.scss';
+
+class JobTile extends Component {
+
+  constructor(props) {
+    super(props);
+    this.state = {}
+  }
+
+  render() {
+    return (
+      <div className="ui link items job-tile">
+        {/* <div className="item">
+          <div className="ui tiny image">
+            <img src="" alt="job tile" />
+          </div>
+          <div className="content">
+            <div className="header">Lead Developer Role</div>
+            <div className="description">
+              <p>Opportunity to join a great team using React and Redux</p>
+            </div>
+          </div>
+        </div> */}
+        <div className="jobs-left">
+          <div className='job-titles'>
+            <h2 className='jobs-title'>Job Title</h2>    
+            <h2 className='job-company-name'>Company Name</h2> 
+          </div>
+
+          <div className='job-desc'>
+            <h2>Do you want to get paid?</h2>
+            <p>This is the coolest job ever. We have snacks and ping pong and we work only when we feel like it</p>
+            <p className='job-posted'>4 Days ago</p>
+          </div>
+        </div>
+        <div className="jobs-right">
+          <div className='job-icons'>
+            <button>&rarr;</button>
+            <button>&#x271A;</button>
+          </div>
+        </div>
+      </div>
+    )
+  }
+}
+
+export default JobTile;
