+++ src/components/JobsContainer/JobTile/JobTile.scss
@@ -1,0 +1,47 @@
+.job-tile {
+    display: flex;
+    width: 400px;
+    border: 3px solid #000;
+
+    .jobs-left {
+        flex-basis: 85%;
+        padding: 15px;
+        .job-titles {
+            display: flex;
+            margin-bottom: 25px;
+            h2 {
+                margin: 0;
+                font-size: 16px;
+                flex-basis: 50%;
+            }
+
+            .job-company-name {
+                text-align: right;
+            }
+        }
+
+        .job-desc {
+            h2 {
+                font-size: 20px;
+            }
+            .job-posted {
+                text-align: right;
+            }
+        }
+    }
+
+    .jobs-right {
+        flex-basis: 15%;
+        border-left: 3px solid #000;
+        .job-icons {
+            display: flex;
+            flex-direction: column;
+            height: 100%;
+            
+            button {
+                height: 50%;
+                cursor: pointer;
+            }
+        }
+    }
+}
